first pass on explore mode pushblocks
