+++ rooms/entry.json
@@ -1,6 +1,6 @@
 {
   "tiles": [
-    "#...#",
+    "....#",
     ".....",
     "..<..",
     ".....",
@@ -11,6 +11,9 @@
     [-1, 2],
     [2, 5],
     [5, 2]
+  ],
+  "elems": [
+    [[1, 1], "PushBlock"]
   ],
   "degree": 1,
   "items": true
