+++ src/contexts/combat.py
@@ -5,7 +5,10 @@ from lib.direction import invert as invert_direction, normal as normalize_direct
 from lib.cell import neighborhood, manhattan, is_adjacent, upscale
 from helpers.combat import find_damage, will_miss, will_crit, will_block
 
-from contexts.explore.base import ExploreBase
+from contexts.explore.base import (
+  ExploreBase,
+  COMMAND_MOVE, COMMAND_MOVE_TO, COMMAND_ATTACK, COMMAND_SKILL, COMMAND_WAIT
+)
 from contexts.skill import SkillContext
 from contexts.gameover import GameOverContext
 from contexts.ally import AllyContext
@@ -30,16 +33,10 @@ from anims.flicker import FlickerAnim
 from vfx.flash import FlashVfx
 from colors.palette import RED, GREEN, BLUE, GOLD, PURPLE, CYAN
 from config import (
-  MOVE_DURATION, FLINCH_PAUSE_DURATION, FLICKER_DURATION, NUDGE_DURATION, PUSH_DURATION, SIDESTEP_DURATION, SIDESTEP_AMPLITUDE,
+  FLINCH_PAUSE_DURATION, FLICKER_DURATION, NUDGE_DURATION, PUSH_DURATION, SIDESTEP_DURATION, SIDESTEP_AMPLITUDE,
   CRIT_MODIFIER,
   TILE_SIZE,
 )
-
-COMMAND_MOVE = "move"
-COMMAND_MOVE_TO = "move_to"
-COMMAND_ATTACK = "attack"
-COMMAND_SKILL = "use_skill"
-COMMAND_WAIT = "wait"
 
 def find_damage_text(damage):
   if damage == 0:
@@ -245,62 +242,12 @@ class CombatContext(ExploreBase):
       ctx.update_bubble()
 
     ctx.hero.facing = delta
-    moved = ctx.move(ctx.hero, delta, on_end=on_move)
+    moved = ctx.move_cell(ctx.hero, delta, on_end=on_move)
     if not moved and issubclass(target_tile, tileset.Pit):
       moved = ctx.leap(actor=ctx.hero, on_end=on_move)
 
     moved and ctx.step()
     return moved
-
-  def move(ctx, actor, delta, duration=0, jump=False, on_end=None):
-    target_cell = vector.add(actor.cell, delta)
-    target_tile = ctx.stage.get_tile_at(target_cell)
-    origin_tile = ctx.stage.get_tile_at(actor.cell)
-    if (not Tile.is_walkable(target_tile)
-    or abs(target_tile.elev - origin_tile.elev) >= 1):
-      return False
-
-    target_elem = (
-      next((e for e in ctx.stage.get_elems_at(target_cell) if e.solid), None)
-      or next((e for e in ctx.stage.get_elems_at(target_cell)), None)
-    )
-
-    if target_elem and target_elem.solid and not (target_elem is ctx.ally and not ctx.ally.command):
-      return False
-
-    move_command = (actor, (COMMAND_MOVE, delta))
-    ctx.command_queue.append(move_command)
-
-    move_duration = duration or MOVE_DURATION
-    move_duration = move_duration * 1.5 if jump else move_duration
-    move_kind = JumpAnim if jump else StepAnim
-    move_anim = move_kind(
-      target=actor,
-      src=actor.cell,
-      dest=target_cell,
-      duration=move_duration,
-      on_end=lambda: (
-        move_command in ctx.command_queue and ctx.command_queue.remove(move_command),
-        on_end and on_end(),
-      )
-    )
-    move_anim.update() # initial update to ensure walk animation loops seamlessly
-
-    move_group = next((g for g in ctx.anims for a in g if isinstance(a, StepAnim) and isinstance(a.target, DungeonActor)), None)
-    not move_group and ctx.anims.append(move_group := [])
-    move_group.append(move_anim)
-    if jump:
-      ctx.anims[-1].append(PauseAnim(duration=move_duration + 5))
-
-    ctx.update_bubble()
-    actor.cell = target_cell
-    actor.facing = normalize_direction(delta)
-    actor.command = move_command
-
-    if target_elem and target_elem is ctx.ally:
-      ctx.move(actor=ctx.ally, delta=invert_direction(delta))
-
-    return True
 
   def move_to(ctx, actor, dest, run=False, on_end=None):
     if actor.cell == dest or actor.ai_path and actor.cell == actor.ai_path[-1]:
@@ -330,7 +277,7 @@ class CombatContext(ExploreBase):
 
     moved = False
     if delta != (0, 0):
-      moved = ctx.move(actor, delta, run, on_end=on_end)
+      moved = ctx.move_cell(actor, delta, run, on_end=on_end)
 
     if not moved:
       actor.ai_path = ctx.stage.pathfind(start=actor.cell, goal=dest)
@@ -380,50 +327,8 @@ class CombatContext(ExploreBase):
 
   def leap(ctx, actor, on_end=None):
     delta = vector.scale(actor.facing, 2)
-    moved = ctx.move(actor, delta, jump=True, on_end=on_end)
+    moved = ctx.move_cell(actor, delta, jump=True, on_end=on_end)
     return moved
-
-  def handle_push(ctx):
-    target_cell = vector.add(ctx.hero.cell, ctx.hero.facing)
-    target_elem = next((e for e in ctx.stage.get_elems_at(target_cell) if e.solid and not e.static), None)
-    if not target_elem:
-      return False
-
-    return ctx.push(
-      actor=ctx.hero,
-      target=target_elem,
-      on_end=lambda: (
-        ctx.update_bubble(),
-        ctx.step(),
-      )
-    )
-
-  def push(ctx, actor, target, on_end=None):
-    src_cell = target.cell
-    dest_cell = vector.add(src_cell, actor.facing)
-    dest_tile = ctx.stage.get_tile_at(dest_cell)
-    dest_elem = next((e for e in ctx.stage.get_elems_at(dest_cell) if e.solid), None)
-    if (target.static
-    or dest_tile is None
-    or dest_tile.solid
-    or dest_tile.pit
-    or dest_elem):
-      return False
-
-    target.cell = dest_cell
-    ctx.move(actor, delta=actor.facing, duration=PUSH_DURATION, on_end=on_end)
-    not ctx.anims and ctx.anims.append([])
-    ctx.anims[-1].extend([
-      StepAnim(
-        target=target,
-        src=src_cell,
-        dest=dest_cell,
-        duration=PUSH_DURATION,
-      ),
-      PauseAnim(duration=15)
-    ])
-    target.on_push(ctx)
-    return True
 
   def handle_action(ctx):
     if not ctx.hero:
@@ -891,9 +796,9 @@ class CombatContext(ExploreBase):
 
     if command_name == COMMAND_MOVE:
       if chain:
-        return ctx.move(actor, *command_args, on_end=on_end)
+        return ctx.move_cell(actor, *command_args, on_end=on_end)
       else:
-        ctx.move(actor, *command_args)
+        ctx.move_cell(actor, *command_args)
         return on_end()
 
     if command_name == COMMAND_MOVE_TO:
